+++ android/domain/src/main/java/com/bonobono/domain/model/registration/Token.kt
@@ -2,7 +2,7 @@ package com.bonobono.domain.model.registration
 
 data class Token(
     val accessToken: String,
-    val accessTokenExpiresIn: Int,
+    val accessTokenExpiresIn: Long,
     val grantType: String,
     val refreshToken: String
 )
